src/recommendation/requirements.txt
pandas
numpy
scikit-learn
tqdm
python-dotenv
google-genai